feat(profile): integrate activity addition with profile management
Added functions in the ProfileViewModel file and the ProfilesRepository files to handle the addition of activities when a user creates or enrolls in an activity. The new `addActivity` method in ProfileViewModel is triggered during activity creation or enrollment to ensure the user's profile is updated accordingly in the Firestore database.