refactor(gallery): update ProfileImage and move it to gallery file
Moved the ProfileImage function to the 'gallery' file and updated its signature to handle URIs directly, supporting the new implementation where images are converted back and forth to strings for database storage. Also, I tried to store the photos as bitmaps, but for the moment, it does not work because bitmaps converted to strings are too big to be stored as a field of an object in firestore, so working on a way to retrieve photos from the database.