feature(Storage) : add functions to upload images of activities and for the profile
Refactored the ProfileImage function to display the profile picture only giving it the user id. Added a method to upload a profile picture in storage, in the user collection. Added a method to upload activity's images in storage, in the activities collection. Each profile or activity has an object with its id as path where their images are stored. Added functions to fetch an activity's images, and a function to fetch a profile's images. Added a gallery screen to pick images from the gallery. Added a carousel to display activities only retrieving them with the activity's id.