refactor(imageRepositoryFirestore): split already existing methods into smaller ones in order to test them in their entirety
Still methods that do not have been splitted such as removeAllActivityImages and fetchImegActivitiesAsBitmaps. For the rest I put the new small methods as public because else I cannot test them and this reduces the coverage a lot. The idea to split methods is because testing addOnClause is really difficult, and I do not think that everything inside can be tested so I took these parts and put them in methods instead.