fix(image activity card): add imageViewModel to signature of ViewListActivities.kt and ActivityCard.
Also add a condition that displays the default picture only if the activity does not have any image. The image is displayed with the modifier to fill the width so that horizontal images fit very well, but this will make the vertical images really bad to use as images for activities.